+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fixed a compilation warning present with Elixir 1.17+ about "`4..1`" not
+  using a range step.
+
 ## 1.0.3 - 2023-04-16
 
 ### Fixed